+++ lib/aixm/component/layer.rb
@@ -95,9 +95,9 @@ module AIXM
       has_one :vertical_limit
 
       # @!method services
-      #   @return [Array<AIXM::Feature::Service>] services
+      #   @return [Array<AIXM::Component::Service>] services
       # @!method add_service(service)
-      #   @param service [AIXM::Feature::Service]
+      #   @param service [AIXM::Component::Service]
       has_many :services
 
       # @!method airspace